HackStudio: Extract and rename the Locator's suggestions model
No code changes besides moving and renaming it.
